# services/api/docs/swagger.yaml
components:
  schemas:
...
    auth.AccessRequestDecideRequest:
      properties:
        approve:
          type: boolean
        note:
          type: string
      type: object
...
    auth.AdminInviteCreateRequest:
      properties:
        days_to_expire:
          type: integer
        email:
          type: string
        email_user:
          type: boolean
        roles:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
    auth.AdminInviteCreateResponse:
      properties:
        expires_at:
          type: string
        invite_id:
          type: string
        invite_link:
          type: string
      type: object
...
    auth.AuditEvent:
      properties:
        actor_id:
          type: string
        created_at:
          type: string
        id:
          type: string
        ip_address:
          type: string
        metadata:
          additionalProperties: {}
          type: object
        type:
          type: string
        user_agent:
          type: string
        user_id:
          type: string
      type: object
    auth.AuditListResponse:
      properties:
        events:
          items:
            $ref: '#/components/schemas/auth.AuditEvent'
          type: array
          uniqueItems: false
        total:
          type: integer
      type: object
...
    auth.CredentialSummary:
      properties:
        aaguid:
          type: string
        device_name:
          type: string
        id:
          type: string
        last_used_at:
          type: string
        sign_count:
          type: integer
      type: object
...
    auth.CredentialsListResponse:
      properties:
        credentials:
          items:
            $ref: '#/components/schemas/auth.CredentialSummary'
          type: array
          uniqueItems: false
      type: object
...
    auth.InvitePreviewResponse:
      properties:
        email:
          type: string
        expires_at:
          type: string
        reason:
          type: string
        valid:
          type: boolean
      type: object
...
    auth.RecoveryGenerateResponse:
      properties:
        codes:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
...
info:
  title: ""
  version: ""
openapi: 3.1.0
paths:
...
  /api/v1/admin/access-requests/{id}:
    patch:
      parameters:
      - description: Access Request ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth.AccessRequestDecideRequest'
        description: Decision
        required: true
      responses:
        "204":
          description: No Content
      summary: Decide access request (admin)
      tags:
      - admin
  /api/v1/admin/audit:
    get:
      parameters:
      - description: Filter by actor ID (UUID)
        in: query
        name: actor_id
        schema:
          type: string
      - description: Filter by user ID (UUID)
        in: query
        name: user_id
        schema:
          type: string
      - description: Comma-separated event types
        in: query
        name: types
        schema:
          type: string
      - description: Filter events created at or after this RFC3339 timestamp
        in: query
        name: since
        schema:
          type: string
      - description: Filter events created before this RFC3339 timestamp
        in: query
        name: until
        schema:
          type: string
      - description: Max results (1-200)
        in: query
        name: limit
        schema:
          default: 50
          type: integer
      - description: Offset for pagination
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth.AuditListResponse'
          description: OK
      summary: List audit events (admin)
      tags:
      - admin
  /api/v1/admin/invites:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth.AdminInviteCreateRequest'
        description: Invite create request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth.AdminInviteCreateResponse'
          description: Created
      summary: Create invite (admin)
      tags:
      - admin
  /api/v1/admin/invites/preview:
    get:
      parameters:
      - description: Invite token (base64url)
        in: query
        name: token
        schema:
          type: string
      - description: Invite ID (UUID)
        in: query
        name: id
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth.InvitePreviewResponse'
          description: OK
      summary: Preview invite (public)
      tags:
      - admin
...
  /api/v1/admin/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
      summary: Delete a user (admin)
      tags:
      - admin
    patch:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "204":
          description: No Content
      summary: Update a user (admin)
      tags:
      - admin
  /api/v1/admin/users/{id}/credentials:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth.CredentialsListResponse'
          description: OK
      summary: List user credentials (admin)
      tags:
      - admin
  /api/v1/admin/users/{id}/recovery/codes/generate:
    post:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth.RecoveryGenerateResponse'
          description: OK
      summary: Generate recovery codes for a user (admin)
      tags:
      - admin
...